--- Proyecto/Segundo Bimestre/boxed-client/src/app/servicios/Auth/auth.service.ts
@@ -1,16 +1,0 @@
-import {Injectable} from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthService {
-  estaLogeado = false;
-  usuario=false;
-  operador=false;
-  constructor() {
-
-  }
-
-
-
-}
